REFACTOR: put tests all back in single file to avoid errors
